Initial
Windows service that will take an address and spit back a lat,long or it
can take an lat,long and spit back an address. I used .NET Remoting for
communicating with the service. The service uses the Google geocoding
API. I used the JsonProvider F# Type Provider to parse the response from
Google.